plugin/tool/pdf_tool/requirements.txt
# Dipendenze specifiche del plugin pdf_tool.
# Installare con: uv pip install -r requirements.txt
agno>=2.5.11
fpdf2>=2.8.0
markdown>=3.5.0
